+++ collector-exoplayer/src/main/java/com/bitmovin/analytics/exoplayer/manipulators/BitrateEventDataManipulator.kt
@@ -6,7 +6,6 @@ import com.google.android.exoplayer2.C.TRACK_TYPE_AUDIO
 import com.google.android.exoplayer2.C.TRACK_TYPE_VIDEO
 import com.google.android.exoplayer2.ExoPlayer
 import com.google.android.exoplayer2.Format
-import com.google.android.exoplayer2.SimpleExoPlayer
 
 class BitrateEventDataManipulator(private val exoplayer: ExoPlayer) : EventDataManipulator {
     var currentAudioFormat: Format? = null
@@ -56,23 +55,22 @@ class BitrateEventDataManipulator(private val exoplayer: ExoPlayer) : EventDataM
     }
 
     fun setFormatsFromPlayer() {
-        currentVideoFormat = (exoplayer as? SimpleExoPlayer)?.videoFormat ?: getCurrentFormatFromPlayer(TRACK_TYPE_VIDEO)
-        currentAudioFormat = (exoplayer as? SimpleExoPlayer)?.audioFormat ?: getCurrentFormatFromPlayer(TRACK_TYPE_AUDIO)
+        currentVideoFormat = exoplayer.videoFormat ?: getCurrentFormatFromPlayer(TRACK_TYPE_VIDEO)
+        currentAudioFormat = exoplayer.audioFormat ?: getCurrentFormatFromPlayer(TRACK_TYPE_AUDIO)
     }
 
     private fun getCurrentFormatFromPlayer(trackType: Int): Format? {
-        if (exoplayer.currentTrackSelections == null) {
+        if (exoplayer.currentTracksInfo == null) {
             return null
         }
 
-        val trackSelection = exoplayer.currentTrackSelections.all
-                .filterIndexed { i, _ -> exoplayer.getRendererType(i) == trackType }
-                .firstOrNull() ?: return null
+        val trackInfo = exoplayer.currentTracksInfo.trackGroupInfos.firstOrNull { track -> track.trackType == trackType }
+                ?: return null
 
-        var format = trackSelection.getFormat(0) ?: null
+        var format = trackInfo.trackGroup.getFormat(0) ?: null
         try {
-            val getSelectedFormatMethod = trackSelection.javaClass.getMethod("getSelectedFormat")
-            format = getSelectedFormatMethod.invoke(trackSelection) as Format
+            val getSelectedFormatMethod = trackInfo.javaClass.getMethod("getSelectedFormat")
+            format = getSelectedFormatMethod.invoke(trackInfo) as Format
         } catch (e: Exception) {
         }